# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.GetUsersResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  handlers.ToggleUserStatusRequest:
    properties:
      is_active:
        type: boolean
    required:
    - is_active
    type: object
...
  handlers.UpdateRoleRequest:
    properties:
      role:
        $ref: '#/definitions/models.Role'
    required:
    - role
    type: object
...
  models.Check:
    properties:
      check_type:
        type: string
      content:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      danger_level:
        type: string
      danger_score:
        type: number
      extracted_text:
        type: string
      file_path:
        type: string
      id:
        type: integer
      processing_time_ms:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Role:
    enum:
    - user
    - moderator
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleModerator
    - RoleAdmin
  models.User:
    properties:
      checks:
        items:
          $ref: '#/definitions/models.Check'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        $ref: '#/definitions/models.Role'
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API для системы детекции скама
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scam Detection API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Возвращает список всех пользователей с пагинацией. Доступно только
        администраторам.
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Количество на странице
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetUsersResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Получить список всех пользователей (только admin)
      tags:
      - admin
  /admin/users/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о пользователе по ID. Доступно
        только администраторам.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Получить информацию о пользователе (только admin)
      tags:
      - admin
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Изменяет роль указанного пользователя. Доступно только администраторам.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Новая роль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Изменить роль пользователя (только admin)
      tags:
      - admin
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      description: Изменяет статус активности пользователя. Доступно только администраторам.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Статус активности
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ToggleUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Заблокировать/разблокировать пользователя (только admin)
      tags:
      - admin
...
securityDefinitions:
  CookieAuth:
    description: JWT токен в HttpOnly cookie
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"